Move board files
