+++ apps/src/lib/kits/maker/boards/circuitPlayground/CircuitPlaygroundBoard.js
@@ -12,7 +12,7 @@ import {
   cleanupCircuitPlaygroundComponents,
   enableCircuitPlaygroundComponents,
   componentConstructors
-} from './PlaygroundComponents';
+} from '../../PlaygroundComponents';
 import {
   SONG_CHARGE,
   SONG_LEVEL_COMPLETE,
@@ -20,15 +20,15 @@ import {
   SONG_CONCLUSION,
   CP_COMMAND,
   J5_CONSTANTS
-} from './PlaygroundConstants';
-import Led from './Led';
+} from '../../PlaygroundConstants';
+import Led from '../../Led';
 import {
   isNodeSerialAvailable,
   ADAFRUIT_VID,
   CIRCUIT_PLAYGROUND_EXPRESS_PID,
   CIRCUIT_PLAYGROUND_PID
-} from './portScanning';
-import {PlaygroundButton} from './Button';
+} from '../../portScanning';
+import {PlaygroundButton} from '../../Button';
 
 // Polyfill node's process.hrtime for the browser, gets used by johnny-five.
 process.hrtime = require('browser-process-hrtime');
