+++ apps/src/lib/kits/maker/toolkit.js
@@ -4,7 +4,7 @@
  */
 import {getStore} from '../../../redux';
 import trackEvent from '../../../util/trackEvent';
-import CircuitPlaygroundBoard from './CircuitPlaygroundBoard';
+import CircuitPlaygroundBoard from './boards/circuitPlayground/CircuitPlaygroundBoard';
 import FakeBoard from './FakeBoard';
 import * as commands from './commands';
 import * as dropletConfig from './dropletConfig';
@@ -16,7 +16,7 @@ import MakerError, {
 import {findPortWithViableDevice} from './portScanning';
 import * as redux from './redux';
 import {isChrome, gtChrome33, isCodeOrgBrowser} from './util/browserChecks';
-import MicroBitBoard from './MicroBitBoard';
+import MicroBitBoard from './boards/microBit/MicroBitBoard';
 import experiments from '@cdo/apps/util/experiments';
 
 // Re-export some modules so consumers only need this 'toolkit' module
