+++ apps/src/lib/kits/maker/util/SetupChecker.js
@@ -1,5 +1,5 @@
 /** @file Stubbable core setup check behavior for the setup page. */
-import CircuitPlaygroundBoard from '../CircuitPlaygroundBoard';
+import CircuitPlaygroundBoard from '../boards/circuitPlayground/CircuitPlaygroundBoard';
 import {ensureAppInstalled, findPortWithViableDevice} from '../portScanning';
 import {
   isCodeOrgBrowser,
